+++ http/src/lib.rs
@@ -169,6 +169,8 @@ pub(crate) use copy::copy;
 pub use error::{Error, Result};
 pub use headers::{HeaderName, HeaderValue, HeaderValues, Headers, KnownHeaderName, SERVER_HEADER};
 pub use http_config::HttpConfig;
+#[cfg(feature = "unstable")]
+pub use http_context::parse_head_for_bench;
 pub use http_context::{HttpContext, run_with_initial_bytes};
 pub use method::Method;
 #[cfg(feature = "unstable")]